requirements.txt
beautifulsoup4==4.12.2
datastructure==0.1.1
fastapi==0.103.2
pydantic==2.4.2
pytest==7.4.2
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==2.0.21
tqdm==4.66.1
uvicorn==0.23.2
mysql-connector-python==8.1.0
httpx==0.25.0
python-jose[cryptography]
python-multipart==0.0.6